huggingface-space/requirements.txt
gradio==4.16.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
onnxruntime==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
